% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverser.R
\name{colin_reverser}
\alias{colin_reverser}
\title{Reverses a string or number}
\usage{
colin_reverser(toReverse)
}
\arguments{
\item{toReverse}{A string or number}
}
\value{
the reverse of the provided string or number
}
\description{
Reverses a string or number
}
\examples{
colin_reverser("foo")
}
